# src/main/resources/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: Genearte swagger to create artifacts for Device API
  version: 1.0.0
  title: Device API
host: virtserver.swaggerhub.com
basePath: /reddy78/device/1.0.0
tags:
- name: Device
  description: Detail device Info.
schemes:
- https
produces:
- application/json
paths:
  /device:
    get:
      tags:
      - Device
      summary: Get all device
      description: Return all devices
      operationId: getDevices
      produces:
      - application/json
      parameters: []
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Device'
        400:
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        404:
          description: Device not found
          schema:
            $ref: '#/definitions/ApiResponse'
        500:
          description: Service unaviable
          schema:
            $ref: '#/definitions/ApiResponse'
  /device/{sku}:
    get:
      tags:
      - Device
      summary: Get device details by sku
      description: Return a single device
      operationId: getDeviceBySKU
      produces:
      - application/json
      parameters:
      - name: sku
        in: path
        description: ID of Device to return
        required: true
        type: string
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Device'
        400:
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        404:
          description: Device not found
          schema:
            $ref: '#/definitions/ApiResponse'
        500:
          description: Service unaviable
          schema:
            $ref: '#/definitions/ApiResponse'
definitions:
  Device:
    type: object
    properties:
      SKU:
        type: string
      Price:
        type: number
        format: float
    description: Model contains device SKU and Price
    example:
      Price: 0.8008282
      SKU: SKU
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string